fix: no broadcast-mode block

+++ integration-test/test.sh
@@ -3,13 +3,17 @@ set -euo pipefail
 IFS=$'\n\t'
 set -x
 
-rm -fr teritori-dapp
-git clone https://github.com/TERITORI/teritori-dapp.git
-cd teritori-dapp
-git checkout 267a3f9604d48c6f1ea5c31d13a5d24c6ed35210
+if [[ -z "${TERITORI_DAPP_REPO:-}" ]]; then
+    rm -fr teritori-dapp
+    git clone https://github.com/TERITORI/teritori-dapp.git
+    cd teritori-dapp
+    git checkout ac44e4e6ee0965d38da5c9736945c9ce08370ea6
+else
+    cd $TERITORI_DAPP_REPO
+fi
 
 yarn
 
-while ! curl http://localhost:1317/node_info; do sleep 1; done
+while ! curl -s http://localhost:26657/status | jq -e '.result.sync_info.latest_block_height|tonumber > 0'; do sleep 5; done
 
 npx tsx packages/scripts/network-setup/deploy teritori-localnet validator
